Rename codec2_interface.* -> freedv_interface.*.
